Do not ignore_missing_imports from mbq packages (#38)
* Do not ignore_missing_imports from mbq packages

This required an upgrade to a typed version of mbq.env, which is a major
version increment. The only breaking change is to mbq.env.get_bool(),
which is not used by this package. Consumers of mbq.metrics and mbq.env,
however, will need to upgrade mbq.env to resolve a dependency conflict.

* Bump to 1.1.6

Consumers of this package and mbq.env will see a dependency conflict.
The solution is to upgrade mbq.env to the latest release.